agents/requirements.txt
web3==6.0.0
openai==1.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.0.3
loguru==0.7.0
tenacity==8.2.2
aiohttp==3.8.5
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0